pyproject.toml
[project]
name = "pysocha"
version = "0.1.0"
description = "Extremely simple static site generator"
authors = [
    { name = "Tod Hansmann", email = "githubpub@todandlorna.com" }
]
dependencies = [
    "markdown-it-py>=3.0.0",
    "jinja2>=3.1.4",
    "typer>=0.12.3",
    "pyyaml>=6.0.1",
    "linkify-it-py>=2.0.3",
    "mdit-py-plugins>=0.4.1",
    "python-frontmatter>=1.1.0",
    "flask>=3.0.3",
    "python-slugify>=8.0.4",
    "python-dateutil>=2.9.0.post0",
]
readme = "README.md"
requires-python = ">= 3.8"

[project.scripts]
pysocha = 'pysocha:main'

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/pysocha"]
